+++ benchmarks/utils.hpp
@@ -17,7 +17,7 @@
 #pragma once
 
 #include <cuco/detail/error.hpp>
-#include <cuco/util/key_generator.hpp>
+#include <cuco/utility/key_generator.hpp>
 
 #include <nvbench/nvbench.cuh>
 
@@ -26,12 +26,12 @@ namespace cuco::benchmark {
 template <typename Dist>
 auto dist_from_state(nvbench::state const& state)
 {
-  if constexpr (std::is_same_v<Dist, cuco::util::dist_type::unique>) {
+  if constexpr (std::is_same_v<Dist, cuco::utility::dist_type::unique>) {
     return Dist{};
-  } else if constexpr (std::is_same_v<Dist, cuco::util::dist_type::uniform>) {
+  } else if constexpr (std::is_same_v<Dist, cuco::utility::dist_type::uniform>) {
     auto const multiplicity = state.get_int64_or_default("Multiplicity", defaults::MULTIPLICITY);
     return Dist{multiplicity};
-  } else if constexpr (std::is_same_v<Dist, cuco::util::dist_type::gaussian>) {
+  } else if constexpr (std::is_same_v<Dist, cuco::utility::dist_type::gaussian>) {
     auto const skew = state.get_float64_or_default("Skew", defaults::SKEW);
     return Dist{skew};
   } else {
@@ -41,6 +41,6 @@ auto dist_from_state(nvbench::state const& state)
 
 }  // namespace cuco::benchmark
 
-NVBENCH_DECLARE_TYPE_STRINGS(cuco::util::dist_type::unique, "UNIQUE", "dist_type::unique");
-NVBENCH_DECLARE_TYPE_STRINGS(cuco::util::dist_type::uniform, "UNIFORM", "dist_type::uniform");
-NVBENCH_DECLARE_TYPE_STRINGS(cuco::util::dist_type::gaussian, "GAUSSIAN", "dist_type::gaussian");
+NVBENCH_DECLARE_TYPE_STRINGS(cuco::utility::dist_type::unique, "UNIQUE", "dist_type::unique");
+NVBENCH_DECLARE_TYPE_STRINGS(cuco::utility::dist_type::uniform, "UNIFORM", "dist_type::uniform");
+NVBENCH_DECLARE_TYPE_STRINGS(cuco::utility::dist_type::gaussian, "GAUSSIAN", "dist_type::gaussian");
